requirements.txt
streamlit
langchain_core
langchain_community
langchain_ollama
langchain_openai
langchain_groq
langchain_google_firestore